requirements.txt
asgiref==3.2.3
Babel==2.8.0
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.2
django-bootstrap==0.2.4
django-bootstrap4==1.1.1
django-heroku==0.3.1
django-phonenumber-field==4.0.0
gunicorn==20.0.4
idna==2.8
phonenumbers==8.11.2
Pillow==7.0.0
psycopg2==2.8.4
python-decouple==3.3
pytz==2019.3
requests==2.22.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.8
whitenoise==5.0.1
